AP-22 Criacao da pagina e da rota de cadastro com ligacao direta ao login
