+++ src/webapp/views/login.html
@@ -18,7 +18,7 @@
                 <div>
                     <input type="password" name="senha" id="senha" placeholder="Digite a sua senha" required="required">
                 </div>
-                <a href="#">Clique aqui para criar a sua conta!</a>
+                <a href="/web/criar-usuario">Clique aqui para criar a sua conta!</a>
                 <button type="submit" class="btn-projetox">Login</button>
             </form>
         </div>
